fix(mdm): harden harness coverage portability
